% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{AAD}
\alias{AAD}
\title{Average Absolute Deviation}
\usage{
AAD(x, i = NULL, na.rm = TRUE, center = NULL)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{center}{numerical value, the central point}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the *average absolute deviation*,
i.e. the mean of the absolute deviations from a central point
(median by default).
}
\examples{
x = c(2,2,3,4,14)
AAD(x = x, center = median(x))#2.8
AAD(x = x, center = mean(x))#3.6
AAD(x = x, center = modalValue(x))#3.0

}
\references{
https://en.wikipedia.org/wiki/Average_absolute_deviation
}
\author{
Alessandro Barberis
}
