% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{areRowValuesEqual}
\alias{areRowValuesEqual}
\title{Check Row Values Are Equal}
\usage{
areRowValuesEqual(x)
}
\arguments{
\item{x}{numerical matrix}
}
\value{
A logical vector, indicating for each row
if there are different elements (\code{FALSE}) or
if the elements are all the same (\code{TRUE}).

If there \code{x} has one column, the function
returns \code{TRUE} for each row.
}
\description{
This function checks each row of the
input matrix to see if there are different elements.
}
\author{
Alessandro Barberis
}
\keyword{internal}
