% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colAADs}
\alias{colAADs}
\title{Column Average Absolute Deviation}
\usage{
colAADs(x, center, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{center}{numerical value, the central point}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[matrixStats]{colMeans2}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the *average absolute deviation*
for each column vector in the input matrix \code{x}.
}
\examples{
x = cbind(c(2,2,3,4,14), c(2,2,3,4,14))
colAADs(x)#2.8 2.8

}
\author{
Alessandro Barberis
}
