% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colArithmeticMeans}
\alias{colArithmeticMeans}
\title{Column Means}
\usage{
colArithmeticMeans(x, na.rm = TRUE, rows = NULL, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{...}{further arguments to \code{\link[matrixStats]{colMeans2}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{mean}
for each column in an input matrix.
}
\details{
It is a wrapper to \code{\link[matrixStats]{colMeans2}}
function.
}
\examples{
x = matrix(data = c(1,3,1,4), ncol = 2)
colArithmeticMeans(x = x)
}
\author{
Alessandro Barberis
}
