% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colEnrichment}
\alias{colEnrichment}
\title{Enrichment Scores}
\usage{
colEnrichment(
  x,
  rows = NULL,
  na.rm = TRUE,
  method = c("ssgsea", "gsva", "plage", "zscore"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{method}{character string, the name of the method
to be used for computing the enrichment scores.
Current available options are:
\describe{
\item{\code{'ssgsea'}}{Single Sample Gene Set Enrichment Analysis}
\item{\code{'gsva'}}{Gene Set Variation Analysis}
\item{\code{'plage'}}{Pathway Level Analysis of Gene Expression}
\item{\code{'zscore'}}{Z-Score}
}}

\item{...}{further arguments to \code{\link[GSVA]{gsva}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes an enrichment score
for each column vector in the input matrix \code{x}.
See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[GSVA]{gsva}}
function.
}
\author{
Alessandro Barberis
}
\keyword{internal}
