% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colIQRs}
\alias{colIQRs}
\title{Column Interquartile Ranges}
\usage{
colIQRs(x, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link{iqr}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the *interquartile range*
for each column vector in the input matrix \code{x}.
See the **Details** section below for further information.
}
\details{
This function applies \code{\link{iqr}}
to each column of the input matrix.
}
\examples{
x = c(7,7,31,31,47,75,87,115,116,119,119,155,177,NA)
x = cbind(x,x)
colnames(x) = c("S1", "S2")
colIQRs(x = x)#88 88

}
\author{
Alessandro Barberis
}
