% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colMADs}
\alias{colMADs}
\title{Column Median Absolute Deviation}
\usage{
colMADs(x, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[matrixStats]{colMads}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{median absolute deviation}
for each column vector in the input matrix \code{x}.
See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[matrixStats]{colMads}}
function.
}
\examples{
x = c(1,1,2,2,4,6,9)
x = cbind(S1 = x, S2 = x)
colMADs(x)#1.4826 1.4826

}
\author{
Alessandro Barberis
}
