% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colMidranges}
\alias{colMidranges}
\title{Column Midranges}
\usage{
colMidranges(x, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{currently not used}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{midrange}
for each column vector in the input matrix \code{x}.
See the \strong{Details} section below for further information.
}
\details{
The midrange score is the average of the lowest
and highest values in a set of data.
It is a measure of central tendency like
the mean, median, and mode. However, it is more prone to bias than
these other measures because it relies solely upon the two most
extreme scores, which could potentially be outliers.

This function calculates the range of values in each
column of \code{x} by calling \code{\link[matrixStats]{colRanges}}.
Then, it uses the min and max values to compute the score.
}
\examples{
x = matrix(data = c(1,2,2,1,1,3), ncol = 2)
colMidranges(x = x)

}
\author{
Alessandro Barberis
}
