% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colModes}
\alias{colModes}
\title{Column Modes}
\usage{
colModes(x, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link{modalValue}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{mode}
for each column vector in the input matrix \code{x}.
}
\examples{
x = matrix(data = c(1,2,2,1,1,3), ncol = 2)
colModes(x = x)

}
\author{
Alessandro Barberis
}
