% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colSsgsea}
\alias{colSsgsea}
\title{Single Sample Gene Set Enrichment Analysis (ssGSEA) Scores}
\usage{
colSsgsea(x, rows = NULL, na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[GSVA]{gsva}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes an enrichment score
for each column vector in the input matrix \code{x}.
See the \strong{Details} section below for further information.
}
\details{
It is a wrapper to \code{\link[GSVA]{gsva}}
function.
}
\author{
Alessandro Barberis
}
