% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colTrimeans}
\alias{colTrimeans}
\title{Column Trimeans}
\usage{
colTrimeans(x, rows = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link{trimean}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{trimean}
for each column vector in the input matrix \code{x}.
See the \strong{Details} section below for further information.
}
\details{
This function applies \code{\link{trimean}}
to each column of the input matrix.
}
\examples{
x = matrix(data = c(1,2,2,1,1,3), ncol = 2)
colTrimeans(x = x)

}
\author{
Alessandro Barberis
}
