% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colTrimmedMeans}
\alias{colTrimmedMeans}
\title{Column Trimmed Means}
\usage{
colTrimmedMeans(x, trim = 0, rows = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link{trimmedMean}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{trimmed mean}
for each column in an input matrix.
}
\examples{
x = matrix(data = sample(seq_len(10)), ncol = 2)
colTrimmedMeans(x = x, trim = 0)
colTrimmedMeans(x = x, trim = 0.2)

}
\author{
Alessandro Barberis
}
