% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-significance.R
\name{computeASL}
\alias{computeASL}
\title{Achieved Significance Level (ASL)}
\usage{
computeASL(obs, rnd, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{obs}{numerical vector containing the observed value of
the test statistic}

\item{rnd}{numerical vector containing the permutation/bootstrap
replication of the statistics}

\item{alternative}{a character string specifying the
alternative hypothesis, must be one of \code{"two.sided"} (default),
\code{"greater"} or \code{"less"}}
}
\value{
A numerical value representing the approximate
achieved significance level (ASL) of the test.
}
\description{
This function computes the approximate
achieved significance level (ASL) for a permutation
test or bootstrap hypothesis test.
}
\details{
It is computed as:

\deqn{\hat{ASL}_{sample}(x) = \frac{1}{n}\sum_{i=1}^{n}  {\hat{\theta}_{i}}^{*} \ge \hat{\theta}}

where \eqn{\hat{\theta}} is a test statistic; \eqn{{\hat{\theta}}^{*}}
is the permutation/bootstrap distribution of \eqn{\hat{\theta}};
\eqn{{\hat{\theta}_{i}}^{*}} is the computed statistic on the
\eqn{i}-th permutation/boostrap vector.

\eqn{\hat{\theta}} is observed, and the ASL of the test
represents the probability of observing at least that large
a value when the null hypothesis is true.
}
\seealso{
Efron, B. and Tibshirani, R.J., An Introduction to the Bootstrap, pp.208 (1994)
}
\author{
Alessandro Barberis
}
\keyword{internal}
