% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-significance.R
\name{computeBcaCI}
\alias{computeBcaCI}
\title{Bias-corrected and Accelerated Confidence Interval}
\usage{
computeBcaCI(...)
}
\description{
Bias-corrected and Accelerated Confidence Interval
}
\keyword{internal}
