% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-significance.R
\name{computeCI}
\alias{computeCI}
\title{Confidence interval of estimate}
\usage{
computeCI(
  obs,
  rnd,
  conf.level = 0.95,
  type = c("standard", "percentile"),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{obs}{a numeric scalar representing the observed value of the summary
statistic. Used only if \code{type = "standard"}}

\item{rnd}{numerical vector containing the permutation/bootstrap
replication of the statistics.}

\item{conf.level}{the desired confidence level.}

\item{type}{character string indicating which formula to use
in the computation of the confidence interval. Available options are:
\describe{
\item{\code{"standard"}}{the standard confidence interval}
\item{\code{"percentile"}}{the percentile confidence interval}
}}

\item{na.rm}{logical, whether NA values should be removed before computation. Defaults to \code{FALSE}.}

\item{...}{further arguments to \code{\link{computeStandardCI}} or \code{\link{computePercentileCI}}}
}
\value{
A named numeric vector with 2 elements, \code{lower} and \code{upper},
the lower and upper bounds of the confidence interval
for the point estimate.
}
\description{
This function computes an approximate confidence interval of
a point estimate (i.e. the test statistic).
}
\seealso{
\code{\link{computeStandardCI}},
\code{\link{computePercentileCI}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
