% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{computeColMeasures}
\alias{computeColMeasures}
\title{Computes Measure for Each Column in a Matrix}
\usage{
computeColMeasures(
  x,
  rows = NULL,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqm", "iqr", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{score}{character string indicating the statistical measure
to compute}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function calculates statistical measure(s)
for each column in an input matrix.
It is an internal function, and it is not intended to be
called by users.
}
\seealso{
\code{\link{colSummations}},
\code{\link{colWeightedSums}},
\code{\link{colArithmeticMeans}},
\code{\link{colTrimmedMeans}},
\code{\link{colWeightedArithmeticMeans}},
\code{\link{colMidpoints}},
\code{\link{colModes}},
\code{\link{colMidranges}},
\code{\link{colMidhinges}},
\code{\link{colTrimeans}},
\code{\link{colIQRs}},
\code{\link{colIQMs}},
\code{\link{colMADs}},
\code{\link{colAADs}},
\code{\link{colSsgsea}},
\code{\link{colGsva}},
\code{\link{colPlage}},
\code{\link{colZscore}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
