% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{computeColScores}
\alias{computeColScores}
\title{Compute Score for Each Column of x}
\usage{
computeColScores(
  x,
  i,
  w,
  na.rm = T,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  transform = c("none", "stepFunction", "quantile"),
  transform.args = list(),
  transform.sub = FALSE,
  ...
)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{w}{numerical vector of weights the same length as
\code{i}. It is used for the computation of the weighted scores}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{score}{character string indicating the summary score to compute}

\item{transform}{string, the transformation method.
Available options are:
\describe{
\item{\code{'none'}}{\code{x} is returned}
\item{\code{'stepFunction'}}{step function}
\item{\code{'quantile'}}{quantile normalisation}
}
See \code{\link{transformData}} for further details}

\item{transform.args}{list of parameters to the
data transformation function}

\item{transform.sub}{logical, whether to transform
\code{x} after it is subset for \code{i} (used to
speedup computation). Default is \code{FALSE},
meaning the transformation would be applied directly
to \code{x} provided in input}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes a summary score for
each column in an input matrix.
}
\author{
Alessandro Barberis
}
\keyword{internal}
