% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{computeMatrixScore}
\alias{computeMatrixScore}
\title{Compute Score}
\usage{
computeMatrixScore(
  x,
  i,
  w,
  na.rm,
  score = c("mean", "trimmedMean", "weightedMean", "median", "mode", "midrange",
    "midhinge", "trimean", "briskow", "reviewedBriskow", "IQM", "weightedSum", "ssGSEA",
    "gsva", "plage", "zscore"),
  ...
)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{w}{numerical vector of weights the same length as
\code{i}}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{score}{character string indicating the summary score to compute}

\item{...}{further arguments to the function used
for computing the scores}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
This function computes one summary
score from each column vector in the input matrix.
}
\author{
Alessandro Barberis
}
\keyword{internal}
