% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{computeMatrixScores}
\alias{computeMatrixScores}
\title{Compute Scores}
\usage{
computeMatrixScores(
  x,
  i,
  na.rm = TRUE,
  scores = c("mean", "trimmedMean", "weightedMean", "median", "mode", "midrange",
    "midhinge", "trimean", "bristow", "reviewedBristow", "IQM", "weightedSum", "ssGSEA",
    "gsva", "plage", "zscore"),
  sample.id = FALSE,
  args,
  cores = 1L
)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the scores}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{scores}{character vector, indicating the
summary score(s) to compute}

\item{sample.id}{logical, whether to report the
sample ID as a column in the output data frame}

\item{args}{named list, where the names must match the
\code{scores}. Each element in the list is another list
containing the arguments to pass to the function used
for computing the named score. For example,
\code{args = list(trimmedMean = list(trim = 0.4))}
indicates to use \code{trim = 0.4} when computing the
trimmed mean scores}

\item{cores}{number of cores to use for parallel execution.}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
TThis function computes summary
score(s) of the signature \code{i} in input
considering each column vector in the input matrix
\code{x}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
