% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{computeMeasure}
\alias{computeMeasure}
\title{Compute Measure}
\usage{
computeMeasure(
  x,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqm", "iqr", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{(named) numerical vector or a matrix features-by-samples}

\item{score}{character string indicating the statistical measure
to compute}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical vector representing the computed measure(s).
}
\description{
This function computes statistical measure(s)
from an input vector or matrix.
It is an internal function, and it is not intended to be
called by users.
}
\details{
If \code{x} is a vector, the output is a single value,
if \code{x} is a matrix, the output is a vector of length
\code{ncol(x)}.

A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
If \code{x} is a matrix, the output is a vector of \code{NA}
values.
}
\seealso{
\code{\link{computeVectorMeasure}},
\code{\link{computeColMeasures}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
