% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{computeScore}
\alias{computeScore}
\title{Compute Score}
\usage{
computeScore(
  x,
  i,
  na.rm,
  score = c("mean", "trimmedMean", "weightedMean", "median", "mode", "midrange",
    "midhinge", "trimean", "bristow", "reviewedBristow", "IQM", "weightedSum", "ssGSEA",
    "gsva", "plage", "zscore"),
  ...
)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{score}{character string indicating the summary score to compute}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes a summary score
from an input vector
}
\author{
Alessandro Barberis
}
\keyword{internal}
