% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{computeSigScoresCorrelation}
\alias{computeSigScoresCorrelation}
\title{Compute Correlation}
\usage{
computeSigScoresCorrelation(
  data,
  adjust = "BH",
  method = c("pearson", "spearman", "kendall"),
  ...
)
}
\arguments{
\item{data}{data frame, output of \code{\link{computeSigScores}}}

\item{adjust}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). See \code{\link{p.adjust}} for details about why to use "holm" rather than "bonferroni"). }

\item{method}{method="pearson" is the default value.  The alternatives to be passed to cor are "spearman" and "kendall".  These last two are much slower, particularly for big data sets. }

\item{...}{further arguments to internal function call}
}
\description{
This function computes the correlation
between summary scores.

It is a wrapper around \code{\link[psych]{corr.test}}
function.
}
\seealso{
\code{\link[psych]{corr.test}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
