% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-significance.R
\name{computeStandardCI}
\alias{computeStandardCI}
\title{Standard Confidence Interval}
\usage{
computeStandardCI(
  obs,
  rnd,
  conf.level = 0.95,
  distribution = c("normal", "t"),
  n,
  na.rm = FALSE
)
}
\arguments{
\item{obs}{a numeric scalar representing the observed value of the summary statistic.}

\item{rnd}{numerical vector containing the permutation/bootstrap
replication of the statistics.}

\item{conf.level}{the desired confidence level.}

\item{distribution}{sampling distribution of the estimate.
Use \code{"normal"} if the population has unknown mean and known variance (or
if the sample size \code{n} is large); \code{"t"} if the population variance is
unknown and sample size is small.}

\item{n}{sample size, used to compute the degrees of freedom if \code{distribution = "t"}.}

\item{na.rm}{logical, whether NA values should be removed before computing the
standard error. Defaults to \code{FALSE}.}
}
\value{
A named numeric vector with 2 elements, \code{lower} and \code{upper},
the lower and upper bounds of the confidence interval
for the point estimate.
}
\description{
This function computes an approximate confidence interval of
a point estimate (i.e. the summary statistic).
}
\details{
The confidence interval provides additional information about the
variability of a point estimate and it is generally defined as

\deqn{CI = estimate \pm \textit{margin of error} = estimate \pm \textit{critical value} \times \textit{standard error}}

where \eqn{\textit{estimate}} is the sample statistic estimating the population parameter of interest;
\eqn{\textit{critical value}} is a value based on the sampling distribution of the estimate
and the desired confidence level; \eqn{\textit{standard error}} is the standard deviation of the
point estimate.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5723800/}
}
\author{
Alessandro Barberis
}
\keyword{internal}
