% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{computeVectorMeasure}
\alias{computeVectorMeasure}
\title{Compute Measure}
\usage{
computeVectorMeasure(
  x,
  i = NULL,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqm", "iqr", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{score}{character string indicating the statistical measure
to compute}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes a statistical measure
from an input vector. It is an internal function, and
it is not intended to be called by users.
}
\seealso{
\code{\link{summation}},
\code{\link{weightedSum}},
\code{\link{arithmeticMean}},
\code{\link{trimmedMean}},
\code{\link{weightedMean}},
\code{\link{midpoint}},
\code{\link{modalValue}},
\code{\link{midrange}},
\code{\link{midhinge}},
\code{\link{trimean}},
\code{\link{iqr}},
\code{\link{IQM}},
\code{\link{MAD}},
\code{\link{AAD}},
\code{\link{ssgsea}},
\code{\link{gsva}},
\code{\link{plage}},
\code{\link{zscore}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
