% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{connectionUtilityFunctions}
\alias{connectionUtilityFunctions}
\alias{filePathExists}
\alias{is.connection}
\alias{checkFilePathAndOpenConnection}
\alias{isOpenConnection}
\alias{checkConAndClose}
\alias{closeCon}
\alias{openCon}
\title{Connection Utility Functions}
\usage{
filePathExists(path = NULL)

is.connection(con)

checkFilePathAndOpenConnection(path = NULL)

isOpenConnection(con)

checkConAndClose(con)

closeCon(object)

openCon(object)
}
\arguments{
\item{path}{character string, the file path}

\item{con}{object to be tested}

\item{object}{a \code{\link{Logger}}}
}
\value{
A logical value, whether the path to the file exists.

Returns \code{TRUE} or \code{FALSE} depending on whether
its argument is of \code{connection} class or not.

An opened connection.

Returns \code{TRUE} or \code{FALSE} depending on whether
its argument is an opened \code{connection}.
}
\description{
Functions to check existence, create, open, and close
connections.
}
\section{Functions}{
\itemize{
\item \code{filePathExists()}: Checks the existence of a file path

\item \code{is.connection()}: Checks if \code{con}
is a connection.

\item \code{checkFilePathAndOpenConnection()}: Checks the existence of a
file path and returns an opened connection.

\item \code{isOpenConnection()}: Returns \code{TRUE} or
\code{FALSE} depending on whether its argument is an opened
\code{connection}.

\item \code{checkConAndClose()}: Closes an
opened connection.

\item \code{closeCon()}: Closes an
opened connection.

\item \code{openCon()}: Opens a
connection.

}}
\author{
Alessandro Barberis
}
\keyword{internal}
