% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{createLogger}
\alias{createLogger}
\title{Logger Class Constructor}
\usage{
createLogger(
  path = character(),
  con = NULL,
  verbose = TRUE,
  level = c("INFO", "DEBUG", "TRACE", "ALL", "OFF")
)
}
\arguments{
\item{path}{a path to the log file to use}

\item{con}{(optional) connection to a file}

\item{verbose}{logical, whether R should report extra information on progress}

\item{level}{the log level of the logger. Five levels are supported:
\describe{
\item{\code{"OFF"}}{indicates the logger is inactive}
\item{\code{"INFO"}}{indicates informational messages}
\item{\code{"DEBUG"}}{indicates fine-grained informational messages}
\item{\code{"TRACE"}}{indicates finer-grained informational events than the \code{"DEBUG"}}
\item{\code{"ALL"}}{all levels are considered}
}
Levels are considered ordered: \code{OFF < INFO < DEBUG < TRACE < ALL}}
}
\value{
A \code{\link{Logger}} object.
}
\description{
Constructor for the object.
}
\details{
A log request in a Logger of \code{level} \strong{l}
is enabled if the level of the request is \strong{<= l}.
}
\seealso{
\code{\link{logAll}},
\code{\link{logTrace}},
\code{\link{logDebug}},
\code{\link{logInfo}}
}
\author{
Alessandro Barberis
}
