% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{dirExists}
\alias{dirExists}
\title{Check the existence of a directory}
\usage{
dirExists(path)
}
\arguments{
\item{path}{character string, a directory path}
}
\value{
\code{TRUE} if \code{output} is a path to a directory,
\code{FALSE} otherwise.
}
\description{
This function checks the existence of a directory in input.
See the \strong{Details} section below for further information.
}
\details{
The function checks that the path exists.
If not, it creates the entire path by using
\code{\link[base]{dir.create}} with \code{recursive = TRUE}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
