% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{genericScorer}
\alias{genericScorer}
\title{Generic Scorer Function}
\arguments{
\item{x}{(named) numerical vector or matrix}

\item{i}{(optional) numerical vector giving the (row) position
in \code{x} or character vector matching the (row) names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{w}{numerical vector of weights the same length as
\code{nrow(x)}. It is used for the computation of the weighted scores}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{score}{character string indicating the summary score to compute}

\item{...}{further arguments to \code{score}}

\item{transform.fun}{function to transform the data.
If provided, \code{x} is transformed
(\code{x = transform.fun(x, transform.args)})
before the computation of the scores.
See \code{\link{getDataTransformer}} for further details
about built-in options}

\item{transform.args}{list of parameters to the
data transformation function}

\item{transform.sub}{logical, whether to transform
\code{x} after it is subset for \code{i} (used to
speedup computation). Default is \code{FALSE},
meaning the transformation would be applied directly
to \code{x} provided in input}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes a summary score
from an input vector or from each column vector
of an input matrix.
}
\seealso{
\code{\link{getDataTransformer}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
