% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{getArgsInFunFormals}
\alias{getArgsInFunFormals}
\title{Check Function Arguments}
\usage{
getArgsInFunFormals(fun, args = NULL, ..., verbose = F)
}
\arguments{
\item{fun}{function}

\item{args}{list of arguments}

\item{...}{arguments passed as \code{name = vale}}
}
\value{
List of arguments that can be used in \code{fun}.
}
\description{
This function check if the arguments \code{args}
are present in the formal arguments of \code{fun}, and returns
only the ones that can be used.
}
\author{
Alessandro Barberis
}
\keyword{internal}
