% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getAvailableDataTransformers}
\alias{getAvailableDataTransformers}
\title{Available Data Transformation Methods}
\usage{
getAvailableDataTransformers()
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the data transformer, to be used in the function calls}
\item{name}{the name of the data transformation method}
}
}
\description{
This function returns the currently available
data transformation methods.
}
\examples{

getAvailableDataTransformers()

}
\author{
Alessandro Barberis
}
