% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getAvailableRandomSamplers}
\alias{getAvailableRandomSamplers}
\title{Available Random Sampling Methods}
\usage{
getAvailableRandomSamplers()
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the random sampling method, to be used in the function calls}
\item{name}{the name of the random sampling method}
}
}
\description{
This function returns the currently available
random sampling methods.
}
\examples{

getAvailableRandomSamplers()

}
\author{
Alessandro Barberis
}
