% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{getAvailableScores}
\alias{getAvailableScores}
\title{Available Summary Scores}
\usage{
getAvailableScores()
}
\value{
A data frame with two columns:

\describe{
\item{id}{the id of the summary score, to be used in the function calls}
\item{name}{the name of the summary score}
}
}
\description{
This function returns the currently available
summary scores.
}
\examples{

getAvailableScores()
}
\author{
Alessandro Barberis
}
