% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{getDataTransformationFunction}
\alias{getDataTransformationFunction}
\title{Get Data Transformation Function}
\usage{
getDataTransformationFunction(f = c("stepFunction", "quantile"))
}
\arguments{
\item{f}{string, the transformation method.
Available options are:
\describe{
\item{\code{'stepFunction'}}{step function}
\item{\code{'quantile'}}{quantile normalisation}
}}
}
\value{
A data transformation function.
}
\description{
This function is a dispatcher for the
selected transformation method.
}
\seealso{
\code{\link{stepFunctionTranformation}},
\code{\link{quantileNormalization}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
