% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{getDataTransformer}
\alias{getDataTransformer}
\title{Get Data Transformation Function}
\usage{
getDataTransformer(f = c("stepFunction", "quantile", "z"))
}
\arguments{
\item{f}{string, the transformation method.
Available options are:
\describe{
\item{\code{'stepFunction'}}{step function}
\item{\code{'quantile'}}{quantile normalisation}
\item{\code{'z'}}{z-standardisation}
}}
}
\value{
A data transformation function.
}
\description{
This function is a dispatcher for the
selected transformation method.
}
\seealso{
\code{\link{stepFunctionTranformation}},
\code{\link{quantileNormalization}},
\code{\link{zStandardisation}}
}
\author{
Alessandro Barberis
}
