% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getDefaultNaValue}
\alias{getDefaultNaValue}
\title{Default NA Value}
\usage{
getDefaultNaValue()
}
\value{
The default \code{NA} value.
}
\description{
This function returns a default choice
for \code{NA} values. It is used to have uniform
behavior.
}
\author{
Alessandro Barberis
}
\keyword{internal}
