% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{getIndex}
\alias{getIndex}
\title{Index of Retained Elements in Signature Vector}
\usage{
getIndex(x)
}
\arguments{
\item{x}{numerical vector, returned from \code{updateSig}}
}
\value{
The \code{index} attribute of \code{x}.
}
\description{
This function returns the \code{index}
attribute of an object returned from \code{updateSig}.
}
\examples{
\dontrun{

#x is a vector
x = setNames(
  object = c(1,2,3,4,5),
  nm = LETTERS[1:5]
)
i = updateSig(x, i = LETTERS[c(1,6,5)])
getIndex(i)

#x is a matrix
x = matrix(
  data = sample(x = 1:100, size = 15),
  nrow = 5,
  ncol = 3,
  dimnames = list(LETTERS[1:5], letters[1:3])
)
i = updateSig(x, i = LETTERS[c(1,6,5)])
getIndex(i)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
