% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{getLogLevels}
\alias{getLogLevels}
\title{Get log levels}
\usage{
getLogLevels(level = c("OFF", "INFO", "DEBUG", "TRACE", "ALL"))
}
\arguments{
\item{level}{character string, a log level}
}
\value{
A vector containing all the levels corresponding to \code{level}
}
\description{
Returns all the log levels included in the selected level
}
\author{
Alessandro Barberis
}
\keyword{internal}
