% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{getLogLine}
\alias{getLogLine}
\title{Log Line}
\usage{
getLogLine(
  line = c("long.line.1", "long.line.2", "long.line.3", "short.line.1", "short.line.2",
    "short.line.3")
)
}
\arguments{
\item{line}{character vector, the type of line to return}
}
\value{
A character vector, a line break.
}
\description{
Returns a pre-defined character vector.
}
\author{
Alessandro Barberis
}
\keyword{internal}
