% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{getPackageName}
\alias{getPackageName}
\title{Package Name}
\usage{
getPackageName()
}
\value{
A string containing the name of the
package.
}
\description{
This function returns the name of
the package.
}
\author{
Alessandro Barberis
}
\keyword{internal}
