% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{getScorer}
\alias{getScorer}
\title{Get Scorer}
\usage{
getScorer(
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore")
)
}
\arguments{
\item{score}{character string, one of the supported measures}
}
\value{
A scoring function:

\describe{
\item{\code{"sum"         }}{\code{\link{sumScorer}}}
\item{\code{"weightedSum" }}{\code{\link{weightedSumScorer}}}
\item{\code{"mean"        }}{\code{\link{meanScorer}}}
\item{\code{"trimmedMean" }}{\code{\link{trimmedMeanScorer}}}
\item{\code{"weightedMean"}}{\code{\link{weightedMeanScorer}}}
\item{\code{"median"      }}{\code{\link{medianScorer}}}
\item{\code{"mode"        }}{\code{\link{modeScorer}}}
\item{\code{"midrange"    }}{\code{\link{midrangeScorer}}}
\item{\code{"midhinge"    }}{\code{\link{midhingeScorer}}}
\item{\code{"trimean"     }}{\code{\link{trimeanScorer}}}
\item{\code{"iqr"         }}{\code{\link{iqrScorer}}}
\item{\code{"iqm"         }}{\code{\link{iqmScorer}}}
\item{\code{"mad"         }}{\code{\link{madScorer}}}
\item{\code{"aad"         }}{\code{\link{aadScorer}}}
\item{\code{"ssgsea"      }}{\code{\link{ssgseaScorer}}}
\item{\code{"gsva"        }}{\code{\link{gsvaScorer}}}
\item{\code{"plage"       }}{\code{\link{plageScorer}}}
\item{\code{"zscore"      }}{\code{\link{zscoreScorer}}}
}
}
\description{
This function is a dispatcher for the score
function selected in input.
}
\details{
The scoring functions are created via the generic scorer
\code{\link{genericScorer}} which handles vector or matrix
input by calling \code{\link{computeColMeasures}}.

Internally, \code{\link{computeColMeasures}} uses these
functions to compute the measures:
\code{\link{colSummations}},
\code{\link{colWeightedSums}},
\code{\link{colArithmeticMeans}},
\code{\link{colTrimmedMeans}},
\code{\link{colWeightedArithmeticMeans}},
\code{\link{colMidpoints}},
\code{\link{colModes}},
\code{\link{colMidranges}},
\code{\link{colMidhinges}},
\code{\link{colTrimeans}},
\code{\link{colIQRs}},
\code{\link{colIQMs}},
\code{\link{colMADs}},
\code{\link{colAADs}},
\code{\link{colSsgsea}},
\code{\link{colGsva}},
\code{\link{colPlage}},
\code{\link{colZscore}}
}
\author{
Alessandro Barberis
}
