% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{getScorers}
\alias{getScorers}
\title{Get Scorers}
\usage{
getScorers(
  scores = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore")
)
}
\arguments{
\item{scores}{character vector, the selected supported measures}
}
\value{
A list containing the scoring functions.
}
\description{
This function is a dispatcher for the score
functions selected in input.
}
\seealso{
The generic scorer:
\code{\link{computeScore}}

And the built-in functions used to compute the measures:
\code{\link{colSummations}},
\code{\link{colWeightedSums}},
\code{\link{colArithmeticMeans}},
\code{\link{colTrimmedMeans}},
\code{\link{colWeightedArithmeticMeans}},
\code{\link{colMidpoints}},
\code{\link{colModes}},
\code{\link{colMidranges}},
\code{\link{colMidhinges}},
\code{\link{colTrimeans}},
\code{\link{colIQRs}},
\code{\link{colIQMs}},
\code{\link{colMADs}},
\code{\link{colAADs}},
\code{\link{colSsgsea}},
\code{\link{colGsva}},
\code{\link{colPlage}},
\code{\link{colZscore}}
}
\author{
Alessandro Barberis
}
