% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{getSigScoresAsMatrix}
\alias{getSigScoresAsMatrix}
\title{Get Signature Scores as Matrix}
\usage{
getSigScoresAsMatrix(data)
}
\arguments{
\item{data}{data frame, output of \code{\link{computeSigScores}}}
}
\value{
A numerical matrix, containing the signature scores.
}
\description{
This function takes a data frame output
of \code{\link{computeSigScores}} and returns a
numerical matrix.
}
\author{
Alessandro Barberis
}
\keyword{internal}
