% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-plot-scores.R
\name{ggPlot}
\alias{ggPlot}
\title{Plot Function}
\usage{
ggPlot(
  data = NULL,
  x,
  y,
  color,
  add.violin = F,
  violin.data = NULL,
  add.boxplot = T,
  boxplot.data = NULL,
  boxplot.width = 0.1,
  add.points = T,
  point.data = NULL,
  point.shape = 16,
  point.jitter.w = 0.2,
  point.jitter.h = NULL,
  labs.title = "Plots",
  labs.x = "x",
  labs.y = "y",
  labs.col = character(),
  axis.text.x.angle = 90,
  axis.text.x.size = 9,
  axis.text.y.angle = 0,
  axis.text.y.size = 9,
  facet.rows = NULL,
  facet.cols = NULL,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} containing the data to plot}

\item{x, y, color}{character string, columns in \code{data}
describing which variables
should be mapped to the aesthetics \code{x}, \code{y}, and \code{color}
in \code{\link[ggplot2]{ggplot}}}

\item{add.violin}{logical, whether to include a violin plot in the figure}

\item{violin.data, boxplot.data, point.data}{\code{data.frame}
containing the data to be displayed in the respective layers.
If \code{NULL}, the data is inherited from \code{data}}

\item{add.boxplot}{logical, whether to include a boxplot in the figure}

\item{boxplot.width}{numeric, the width of boxplot}

\item{add.points}{logical, whether to add the computed scores
of the individual summary measures as points in the plot}

\item{point.shape}{the shape to use to plot the scores. It can take
five types of values:
\itemize{
\item An integer in \verb{[0, 25]}
\item The name of the shape
\item A single character, used as a plotting symbol
\item A . to draw the smallest rectangle that is visible,
usually 1 pixel
\item An \code{NA}, to draw nothing
}

See \code{vignette("ggplot2-specs")} for further details}

\item{point.jitter.w, point.jitter.h}{Amount of vertical and
horizontal jitter. The jitter is added in both positive and negative
directions, so the total spread is twice the value specified here.
See \code{\link[ggplot2]{position_jitter}} for further details}

\item{labs.title}{The text for the title}

\item{labs.x}{The title of the x axis}

\item{labs.y}{The title of the y axis}

\item{labs.col}{The title of the legend}

\item{axis.text.x.angle, axis.text.y.angle}{Specify the x and y
axis tick labels angles (in \code{[0, 360]})
See \code{\link[ggplot2]{element_text}} for further details}

\item{axis.text.x.size, axis.text.y.size}{Specify the x and y
axis tick labels size in pts.
See \code{\link[ggplot2]{element_text}} for further details}

\item{facet.rows, facet.cols}{character string, variables
defining the faceting groups on the rows or columns
dimension.
See \code{\link[ggplot2]{facet_grid}} for further details}

\item{...}{further arguments to internal
function call}
}
\value{
A \code{ggplot} object.
}
\description{
This function generates a
\code{ggplot} object.

Inside \code{ggPlot} there are different function calls
to \code{ggplot2} functions in order to create a pre-defined
plot.
}
\author{
Alessandro Barberis
}
\keyword{internal}
