% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{gsvaScores}
\alias{gsvaScores}
\title{Gene Set Variation Analysis (GSVA) Scores}
\usage{
gsvaScores(
  x,
  i,
  na.rm = TRUE,
  verbose = FALSE,
  kcdf = c("Gaussian", "Poisson", "none"),
  ...
)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{verbose}{Gives information about each calculation step. Default: \code{FALSE}.}

\item{kcdf}{Character string denoting the kernel to use during the non-parametric estimation of the
              cumulative distribution function of expression levels across samples when \code{method="gsva"}.
              By default, \code{kcdf="Gaussian"} which is suitable when input expression values are continuous,
              such as microarray fluorescent units in logarithmic scale, RNA-seq log-CPMs, log-RPKMs or log-TPMs.
              When input expression values are integer counts, such as those derived from RNA-seq experiments,
              then this argument should be set to \code{kcdf="Poisson"}.}

\item{...}{further arguments to the function used
for computing the scores}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
Gene Set Variation Analysis (GSVA) Scores
}
\details{
This function uses internally \code{\link[GSVA]{gsva}}
from \code{GSVA} R package. To avoid runtime
errors, if \code{x} has less than 2 rows the
function returns \code{NA}.
}
\seealso{
\code{\link[GSVA]{gsva}}
}
\author{
Alessandro Barberis
}
