% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-plot-scores.R
\name{heatMap}
\alias{heatMap}
\title{Heatmap}
\usage{
heatMap(
  data,
  na_col = "grey",
  color_space = "LAB",
  rect_gp = grid::gpar(col = "white"),
  col = NULL,
  column_title = character(0),
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  ...
)
}
\arguments{
\item{data}{data frame, output of \code{\link{computeSigScores}}}

\item{na_col}{Color for \code{NA} values.}

\item{color_space}{The color space in which colors are interpolated. Only used if \code{matrix} is numeric and  \code{col} is a vector of colors. Pass to \code{\link[circlize]{colorRamp2}}.}

\item{rect_gp}{Graphic parameters for drawing rectangles (for heatmap body). The value should be specified by \code{\link[grid]{gpar}} and \code{fill} parameter is ignored.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{column_title}{Title on the column.}

\item{cluster_rows}{If the value is a logical, it controls whether to make cluster on rows. The value can also be a \code{\link[stats]{hclust}} or a \code{\link[stats]{dendrogram}} which already contains clustering. Check \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#clustering} .}

\item{cluster_columns}{Whether make cluster on columns? Same settings as \code{cluster_rows}.}

\item{...}{further arguments to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A \code{ComplexHeatmap} object.
}
\description{
This function generates a
\code{ComplexHeatmap} object.

It is a wrapper around the \code{\link[ComplexHeatmap]{Heatmap}}
function.
}
\author{
Alessandro Barberis
}
\keyword{internal}
