% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-plot-scores.R
\name{heatmapCorSigScores}
\alias{heatmapCorSigScores}
\title{Plot Correlation of Summary Scores as Heatmap}
\usage{
heatmapCorSigScores(
  data,
  cor = NULL,
  scores = NULL,
  name = "test\\nstatistic",
  column_title = "Correlation",
  col = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  ...
)
}
\arguments{
\item{data}{data frame, output of \code{\link{computeSigScores}}}

\item{cor}{(optional) matrix containing the data to plot.
If missing, a correlation matrix is internally computed using
the default settings of function \code{\link{computeSigScoresCorrelation}}}

\item{scores}{(optional) character vector, indicating
the summary score(s) to plot from \code{data}}

\item{...}{further arguments to internal
function call}
}
\value{
A \code{ComplexHeatmap} object.
}
\description{
This function generates a
\code{ComplexHeatmap} object.
}
\examples{
#Set seed for reproducibility
set.seed(seed = 5381L)

#Define row/col size
n = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = n*n, min = 0, max = 1000),
 nrow = n,
 ncol = n,
 dimnames = list(
    paste0("g",seq(n)),
    paste0("S",seq(n))
 )
)

#Compute 5 Summary Scores
x = computeSigScores(
 x = x,
 i = rownames(x),
 scores = c("mean", "median", "mode", "midrange", "midhinge")
)

#Plot correlation (correlation is internally computed)
heatmapCorSigScores(data = x)




#Use simulated correlation matrix

#Define row/col size
n = 5

#Create matrix
x = matrix(
 data = 1,
 nrow = n,
 ncol = n,
 dimnames = list(
   getAvailableScores()$id[1:n],
   getAvailableScores()$id[1:n]
 )
)

#Create data
data = stats::runif(n = (((n*n)-n) / 2), min = -1, max = 1)

#Update upper triangular
x[upper.tri(x = x, diag = F)] = data

#Update lower triangular to be symmetric
x[lower.tri(x = x, diag = F)] = t(x)[lower.tri(x)]

#Plot
heatmapCorSigScores(cor = x)

}
\seealso{
\code{\link{computeSigScoresCorrelation}},
\code{\link{heatmapSigScores}}
}
\author{
Alessandro Barberis
}
