% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{interquartileMeanScore}
\alias{interquartileMeanScore}
\title{Interquartile Mean (IQM) Score}
\usage{
interquartileMeanScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{na.rm}{unused argument, provided for consistency
with other functions.}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
The interquartile mean is a statistical measure
of central tendency based on the truncated mean of the
interquartile range. It is computed as:

\deqn{IQM(x) = \frac{2}{n}\sum_{i=\frac{n}{4}+1}^{\frac{3n}{4}}  x_{i}}

where \eqn{x_{i}} is the \eqn{i}-th element of the ordered vector.
}
\references{
See https://en.wikipedia.org/wiki/Interquartile_mean
}
\author{
Alessandro Barberis
}
