% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{iqr}
\alias{iqr}
\title{Interquartile Range}
\usage{
iqr(x, i = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the \emph{interquartile range}
of the \code{x} values.
See the \strong{Details} section below for further information.
}
\details{
The \emph{interquartile range} (also called \emph{midspread}
or \emph{H-spread}) is a measure of statistical dispersion.
It is defined as the difference between the 75th and 25th
percentiles of the data:

\deqn{IQR(x) = Q_{3}(x) - Q_{1}(x)}

Since missing values and NaN's are allowed only
if \code{na.rm = TRUE} in \code{\link[stats]{quantile}},
if \code{x} has \code{NA} values and \code{na.rm = FALSE}
then \code{NA} is returned.
}
\examples{
x = c(7,7,31,31,47,75,87,115,116,119,119,155,177,NA)
iqr(x = x)#88

}
\references{
https://en.wikipedia.org/wiki/Interquartile_range
}
\author{
Alessandro Barberis
}
