% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{logDebug}
\alias{logDebug}
\title{Log Method}
\usage{
logDebug(object, message, sep, add.level = FALSE, add.time = FALSE)
}
\arguments{
\item{object}{an object of class \code{\link{Logger}}}

\item{message}{a character string, the message to print}

\item{sep}{a character vector of strings to append after \code{message}}

\item{add.level}{logical, whether to add the log level to \code{message}}

\item{add.time}{logical, whether to add the time to \code{message}}
}
\value{
None (invisible \code{NULL}).
}
\description{
This function prints the log information of level \code{DEBUG}.
}
\details{
A log information of level \code{DEBUG} is printed if the provided logger
has a >= log level.
}
\seealso{
\code{\link{logAll}},
\code{\link{logTrace}},
\code{\link{logInfo}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
