% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{logDefault}
\alias{logDefault}
\title{Log Method}
\usage{
logDefault(
  object,
  log.level,
  message,
  sep,
  add.level = FALSE,
  add.time = FALSE
)
}
\arguments{
\item{object}{an object of class \code{\link{Logger}}}

\item{log.level}{a character string, the log level of the information}

\item{message}{a character string, the message to print}

\item{sep}{a character vector of strings to append after \code{message}}

\item{add.level}{logical, whether to add the log level to \code{message}}

\item{add.time}{logical, whether to add the time to \code{message}}
}
\value{
None (invisible \code{NULL}).
}
\description{
Generic function to print the log information.
}
\seealso{
\code{\link{logAll}},
\code{\link{logTrace}},
\code{\link{logDebug}},
\code{\link{logInfo}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
