% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{Logger}
\alias{Logger}
\title{Logger Structure}
\usage{
Logger(path, con, verbose, level)
}
\arguments{
\item{path}{a path to the log file to use}

\item{con}{(optional) connection to a file}

\item{verbose}{logical, whether R should report extra information on progress}

\item{level}{the log level of the logger. Five levels are supported:
\describe{
\item{\code{"OFF"}}{indicates the logger is inactive}
\item{\code{"INFO"}}{indicates informational messages}
\item{\code{"DEBUG"}}{indicates fine-grained informational messages}
\item{\code{"TRACE"}}{indicates finer-grained informational events than the \code{"DEBUG"}}
\item{\code{"ALL"}}{all levels are considered}
}
Levels are considered ordered: \code{OFF < INFO < DEBUG < TRACE < ALL}}
}
\value{
A Logger object.
}
\description{
The Logger class serves as the core element for the
logging system in \code{sigscores}.
}
\author{
Alessandro Barberis
}
\keyword{internal}
