% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{loggerGettersAndSetters}
\alias{loggerGettersAndSetters}
\alias{getCon}
\alias{getVerbose}
\alias{getPath}
\alias{getLevel}
\alias{setCon}
\alias{setVerbose}
\alias{setPath}
\alias{setLevel}
\title{Logger Getters and Setters}
\usage{
getCon(object)

getVerbose(object)

getPath(object)

getLevel(object)

setCon(object, value)

setVerbose(object, value)

setPath(object, value)

setLevel(object, value)
}
\arguments{
\item{object}{a \code{\link{Logger}}}
}
\description{
Logger Getters and Setters
}
\section{Functions}{
\itemize{
\item \code{getCon()}: Returns the object element

\item \code{getVerbose()}: Returns the object element

\item \code{getPath()}: Returns the object element

\item \code{getLevel()}: Returns the object element

\item \code{setCon()}: Returns the updated object

\item \code{setVerbose()}: Returns the updated object

\item \code{setPath()}: Returns the updated object

\item \code{setLevel()}: Returns the updated object

}}
\author{
Alessandro Barberis
}
\keyword{internal}
