% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{logLevelNameFixedLength}
\alias{logLevelNameFixedLength}
\title{Get log levels}
\usage{
logLevelNameFixedLength(level)
}
\arguments{
\item{level}{character string, a log level}
}
\value{
A vector containing all the levels corresponding to \code{level}
}
\description{
Returns the log \code{level} as a string with fixed length.
}
\author{
Alessandro Barberis
}
\keyword{internal}
