% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{MAD}
\alias{MAD}
\title{Median Absolute Deviation}
\usage{
MAD(x, i = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[stats]{mad}}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the *median absolute deviation*.
}
\details{
It is a wrapper to \code{\link[stats]{mad}}
function.
}
\examples{
x = c(1,1,2,2,4,6,9)
MAD(x)#1.4826

}
\references{
https://en.wikipedia.org/wiki/Median_absolute_deviation
}
\author{
Alessandro Barberis
}
