% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{meanScores}
\alias{meanScores}
\title{Arithmetic Mean Scores}
\usage{
meanScores(x, i = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{...}{further arguments to \code{\link[matrixStats]{colMeans2}}}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
Arithmetic Mean Scores
}
\seealso{
\code{\link[matrixStats]{colMeans2}}
}
\author{
Alessandro Barberis
}
