% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{medianScore}
\alias{medianScore}
\title{Median Scores}
\usage{
medianScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
The \emph{median} is the value separating the lower half
of a set of measures from the higher half.
}
\seealso{
\code{\link[stats]{median}}
}
\author{
Alessandro Barberis
}
