% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{medianScores}
\alias{medianScores}
\title{Median Scores}
\usage{
medianScores(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
This function computes the sample median
from each column vector in the input matrix.
}
\seealso{
\code{\link{medianScore}}
}
\author{
Alessandro Barberis
}
