% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{midhinge}
\alias{midhinge}
\title{Midhinge}
\usage{
midhinge(x, i = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
The *midhinge* of a set of values
is the mean of the first and third quartiles.
See the **Details** section below for further information.
}
\details{
This measure of location is computed as:

\deqn{MH(x) = \frac{Q_{1}(x) + Q_{3}(x)}{2} = \frac{median(x) + midhinge(x)}{2}}

It is related to the interquartile range (\eqn{IQR = Q_{3} - Q_{1}}),
which is a measure of dispersion.

Since missing values and NaN's are allowed only
if \code{na.rm = TRUE} in \code{\link[stats]{quantile}},
if \code{x} has \code{NA} values and \code{na.rm = FALSE}
then \code{NA} is returned.
}
\examples{
x = c(0,1,1,2,3,3,3,100,NA)
midhinge(x = x, na.rm = TRUE)

}
\references{
https://en.wikipedia.org/wiki/Midhinge
}
\author{
Alessandro Barberis
}
