% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{midhingeScore}
\alias{midhingeScore}
\title{Midhinge Score}
\usage{
midhingeScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{unused argument, provided for consistency
with other functions.
Missing values and NaN's are allowed
only if \code{na.rm = TRUE} in \code{\link[stats]{quantile}}.}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
The \emph{midhinge} of a set of values
is the mean of the first and third quartiles.
}
\author{
Alessandro Barberis
}
