% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{midpoint}
\alias{midpoint}
\title{Median}
\usage{
midpoint(x, i = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[stats]{median}}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the \emph{median}.
}
\details{
It is a wrapper to \code{\link[stats]{median}}
function.
}
\examples{
x = c(seq_len(length.out = 10), NA)
midpoint(x = x, na.rm = TRUE)

}
\author{
Alessandro Barberis
}
