% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{midrange}
\alias{midrange}
\title{Midrange}
\usage{
midrange(x, i = NULL, na.rm = T)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the *midrange* score, i.e.
the average of the lowest and highest values in a set of data.
See the **Details** section below for further information.
}
\details{
The midrange score is a measure of central tendency like
the mean, median, and mode. However, it is more prone to bias than
these other measures because it relies solely upon the two most
extreme scores, which could potentially be outliers.
}
\examples{
x = c(0,1,1,2,3,3,3,100,NA)
midrange(x = x, na.rm = TRUE)

}
\references{
https://dictionary.apa.org/midrange
}
\author{
Alessandro Barberis
}
