% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{midrangeScore}
\alias{midrangeScore}
\title{Midrange Score}
\usage{
midrangeScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes the \emph{midrange} score, i.e.
the average of the lowest and highest values in a set of data.
}
\details{
The midrange score is a measure of central tendency like
the mean, median, and mode. However, it is more prone to bias than
these other measures because it relies solely upon the two most
extreme scores, which could potentially be outliers.
}
\references{
https://dictionary.apa.org/midrange
}
\seealso{
\code{\link{meanScore}},
\code{\link{medianScore}},
\code{\link{modeScore}}
}
\author{
Alessandro Barberis
}
