% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{modalValue}
\alias{modalValue}
\title{Mode}
\usage{
modalValue(x, i = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A length-one numerical vector.
}
\description{
Compute the *mode* of a vector, i.e. the value
that has highest number of occurrences. If different values have
the same number of occurrences, the first one is reported.
}
\examples{
x = c(1,1,2,3,3,3,NA)
modalValue(x = x, na.rm = TRUE)

}
\author{
Alessandro Barberis
}
