% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{modeScore}
\alias{modeScore}
\title{Mode Score}
\usage{
modeScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
Compute the \emph{mode} of a vector, i.e. the value
that has highest number of occurrences. If different values have
the same number of occurrences, the first one is reported.
}
\author{
Alessandro Barberis
}
