% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{normaliseData}
\alias{normaliseData}
\title{Normalise Data}
\usage{
normaliseData(x, method = c("quantile"), ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{method}{string, the normalisation method.
Available options are:
\describe{
\item{quantile}{quantile normalisation}
}}

\item{...}{further arguments to the normalisation method}
}
\value{
A numerical matrix containing the normalized
values.
}
\description{
This function is a dispatcher for the
selected normalisation method.
}
\seealso{
\code{\link{quantileNormalization}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
