% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{plageScores}
\alias{plageScores}
\title{Pathway Level Analysis of Gene Expression (PLAGE) Scores}
\usage{
plageScores(x, i, na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{...}{further arguments to the function used
for computing the scores}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
This function computes the PLAGE score
from each column vector in the input matrix.
}
\details{
This function uses internally \code{\link[GSVA]{gsva}}
from \code{GSVA} R package. To avoid runtime
errors, if \code{x} has less than 2 rows the
function returns \code{NA}.
}
\seealso{
\code{\link[GSVA]{gsva}}
}
\author{
Alessandro Barberis
}
