% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-log-functions.R
\name{printLog}
\alias{printLog}
\alias{printLogLine}
\title{Logger Print Functions}
\usage{
printLog(object, message, sep)

printLogLine(
  object,
  line = c("long.line.1", "long.line.2", "long.line.3", "short.line.1", "short.line.2",
    "short.line.3"),
  sep
)
}
\arguments{
\item{object}{a \code{\link{Logger}}}

\item{message}{character string, the message to print}

\item{sep}{a character vector of strings to append after each element.}

\item{line}{character vector, the type of line to return}
}
\description{
This function prints the provided \code{message}
to the output defined via \code{object}.
}
\section{Functions}{
\itemize{
\item \code{printLogLine()}: Prints a log line.

}}
\author{
Alessandro Barberis
}
\keyword{internal}
