% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{quantileNormalizationForVector}
\alias{quantileNormalizationForVector}
\title{Quantile Normalization}
\usage{
quantileNormalizationForVector(x)
}
\arguments{
\item{x}{a (named) numerical vector}
}
\value{
This function always returns \code{x}.
}
\description{
Currently this function just returns
\code{x} as it is.
}
\author{
Alessandro Barberis
}
\keyword{internal}
