% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{randomSignatures}
\alias{randomSignatures}
\title{Random Sampling}
\usage{
randomSignatures(x, i, n = 1L, exclude = T)
}
\arguments{
\item{x}{a numerical vector or matrix}

\item{i}{(optional) numerical vector giving the (row)
elements in \code{x} or character vector matching the
(row) names in \code{x}}

\item{n}{integer, number of repeated samples
to generate}

\item{exclude}{logical, whether to exclude
\code{i} from the possible values}
}
\value{
A list of length \code{n} where each element
is a vector of indices representing the sampled data.
}
\description{
This function takes \code{n} samples of
size \code{nrow(x)} from the elements of \code{x}.
Two types of sampling are available: sampling with
or without replacement.
}
\seealso{
\code{\link[base]{sample}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
