% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{reviewedBristowScore}
\alias{reviewedBristowScore}
\title{Reviewd Bristow Score}
\usage{
reviewedBristowScore(x, i, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes the \emph{reviewed Bristow score}.
See the \strong{Details} section below for further information.
}
\details{
The score is computed in 3 steps:
\enumerate{
\item compute the median of elements in \code{x}
\item if element in \code{i} is greater than the median,
assign a \code{+1} to the score; assign \code{-1} otherwise
\item divide the score by the number of elements in \code{i}
}
}
\author{
Alessandro Barberis
}
