% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-plot-scores.R
\name{scatterplotSigScores}
\alias{scatterplotSigScores}
\title{Plot Summary Scores as Scatter Plots}
\usage{
scatterplotSigScores(
  data,
  scores = NULL,
  runs = NULL,
  add.scores = T,
  scores.shape = 16,
  scores.jitter.w = 0.2,
  scores.jitter.h = NULL,
  labs.title = "Summary Scores",
  labs.x = "Sample IDs",
  labs.y = "Scores",
  labs.col = "Summary\\nScores",
  axis.text.x.angle = 90,
  axis.text.x.size = 9,
  axis.text.y.angle = 0,
  axis.text.y.size = 9,
  ...
)
}
\arguments{
\item{data}{data frame, output of \code{\link{computeSigScores}}}

\item{scores}{(optional) character vector, indicating
the summary score(s) to plot from \code{data}}

\item{runs}{(optional) numeric vector, indicating the
repeats to plot from \code{data}}

\item{labs.title}{The text for the title}

\item{labs.x}{The title of the x axis}

\item{labs.y}{The title of the y axis}

\item{labs.col}{The title of the legend}

\item{axis.text.x.angle, axis.text.y.angle}{Specify the x and y
axis tick labels angles (in [0, 360])
See \code{\link[ggplot2]{element_text}} for further details}

\item{axis.text.x.size, axis.text.y.size}{Specify the x and y
axis tick labels size in pts.
See \code{\link[ggplot2]{element_text}} for further details}

\item{...}{further arguments to \code{\link{ggPlot}}}
}
\value{
A \code{ggplot} object.
}
\description{
This function generates a
\code{ggplot} object.

\code{scatterplotSigScores} uses the internal function
\code{\link{ggPlot}} which contains different function
calls to \code{ggplot2} functions in order to create a
pre-defined plot.
}
\examples{
#Set seed
set.seed(seed = 5381L)

#Define row/col size
n = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = n*n, min = 0, max = 100),
 nrow = n,
 ncol = n,
 dimnames = list(
    paste0("g",seq(n)),
    paste0("S",seq(n))
 )
)

#Compute Summary Scores
x = computeSigScores(
 x = x,
 i = rownames(x)
)

#Plot scores
scatterplotSigScores(data = x)

#Plot scatter plots per summary score
scatterplotSigScores(
 data       = x,
 scores     = c("mean", "median"),
 facet.rows = "summaryScore"
)
}
\seealso{
\code{\link{ggPlot}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
