% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{stepFunction}
\alias{stepFunction}
\alias{stepFunction.vector}
\alias{stepFunction.matrix}
\title{Step Function}
\usage{
stepFunction(x, ...)

\method{stepFunction}{vector}(x, y = c(-1, 0, 1), thr = 0)

\method{stepFunction}{matrix}(x, y = c(-1, 0, 1), thr = 0)
}
\arguments{
\item{x}{a (named) numerical vector or a matrix features-by-samples}

\item{...}{further arguments passed to or from other methods}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical value, the threshold to use to
divide values in \code{x} in two intervals}
}
\value{
A numerical vector or matrix, containing the output
of the step function for each element of \code{x}.
The default settings of \code{thr} and \code{y}
makes this function equivalent to a sign function.
}
\description{
This function implements a \emph{Step Function}.
See the \strong{Details} section below for further information.
}
\details{
This function implements a piecewise function
that is defined as follows:

\deqn{SF = f(x) =
\left\{
\begin{array}{l}
y_{1}, \quad \textrm{if} \quad x > threshold \\
y_{2}, \quad \textrm{if} \quad x = threshold \\
y_{3}, \quad \textrm{if} \quad x < threshold
\end{array}
\right\}
}
}
\section{Methods (by class)}{
\itemize{
\item \code{stepFunction(vector)}: returns a vector with the transformed values

\item \code{stepFunction(matrix)}: returns a matrix with transformed values

}}
\author{
Alessandro Barberis
}
\keyword{internal}
