% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{stepFunctionForMatrix}
\alias{stepFunctionForMatrix}
\title{Step Function}
\usage{
stepFunctionForMatrix(x, y = c(-1, 0, 1), thr = 0, by = c("rows", "cols"))
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical vector, the threshold(s) to use to
divide values in \code{x} in two intervals.
If it is a single value, the same threshold is applied
throughout \code{x}.
If \code{by = 'rows'} and \code{thr} has number
of elements equivalent to rows in \code{x} then
a different threshold is applied to each row.
If \code{by = 'cols'} and \code{thr} has number
of elements equivalent to columns in \code{x} then
a different threshold is applied to each column}

\item{by}{character string, indicating whether to
apply the threshold to rows or columns of
\code{x}.
In the first case, the threshold for the \eqn{i}-th row
is the same across columns.
In the second case, the threshold is the same within
the same column vector}
}
\value{
A numerical matrix, containing the output
of the step function for each element of \code{x}.
The default settings of \code{thr} and \code{y}
makes this function equivalent to a sign function.
}
\description{
This function implements a \emph{Step Function}.
See the \strong{Details} section below for further information.
}
\details{
This function implements a piecewise function
that is defined as follows:

\deqn{SF = f(x) =
\left\{
\begin{array}{l}
y_{1}, \quad \textrm{if} \quad x > threshold \\
y_{2}, \quad \textrm{if} \quad x = threshold \\
y_{3}, \quad \textrm{if} \quad x < threshold
\end{array}
\right\}
}
}
\author{
Alessandro Barberis
}
\keyword{internal}
