% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{stepFunctionTranformation}
\alias{stepFunctionTranformation}
\title{Step Function Transformation}
\usage{
stepFunctionTranformation(
  x,
  y = c(-1, 0, 1),
  thr = NULL,
  method = c("median", "mean", "mode", "midrange", "trimean", "iqm", "iqr", "mad"),
  by = c("rows", "cols"),
  na.rm = TRUE
)
}
\arguments{
\item{x}{a (named) numerical vector or a matrix features-by-samples}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical value, the threshold to use to
divide values in \code{x} in two intervals}

\item{method}{character string, indicating the score to be used as
threshold. This parameter is only used when \code{thr = NULL} or
not provided}

\item{by}{character string, indicating whether to compute the
threshold by applying \code{method} to rows or columns of
\code{x}.
In the first case, the threshold for the \eqn{i}-th row
is the same across columns.
In the second case, the threshold is the same within
the same column vector.
It is used when \code{x} is a matrix}

\item{na.rm}{logical, whether to remove \code{NA} values
before the computation of the threshold}
}
\value{
A numerical vector or matrix, containing the output
of the transformation.
}
\description{
This function transform input data via
a \emph{Step Function}.
See \code{\link{stepFunction}} for further information.
}
\examples{
#set seed for reproducibility
set.seed(seed = 5381L)

#x is a vector
x = sample(x = -10:10, size = 10)
stepFunctionTranformation(x, thr = 0)

#x is a matrix
#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)
#compute
stepFunctionTranformation(x, thr = 500)

#use median as threshold
stepFunctionTranformation(x, method = 'median')

#use median as threshold (by columns)
stepFunctionTranformation(x, method = 'median', by = 'cols')

}
\author{
Alessandro Barberis
}
