% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{stepFunctionTranformationForMatrix}
\alias{stepFunctionTranformationForMatrix}
\title{Step Function Transformation for Matrix}
\usage{
stepFunctionTranformationForMatrix(
  x,
  y = c(-1, 0, 1),
  thr,
  method = c("median", "mean", "mode", "midrange", "trimean", "iqm", "iqr", "mad"),
  by = c("rows", "cols"),
  na.rm = T
)
}
\arguments{
\item{x}{a matrix features-by-samples}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical value, the threshold to use to
divide values in \code{x} in two intervals}

\item{method}{character string, indicating the score to be used as
threshold. This parameter is only used when \code{thr = NULL} or
not provided}

\item{by}{character string, indicating whether to compute the
threshold by applying \code{method} to rows or columns of
\code{x}.
In the first case, the threshold for the \eqn{i}-th row
is the same across columns.
In the second case, the threshold is the same within
the same column vector.
It is used when \code{x} is a matrix}

\item{na.rm}{logical, whether to remove \code{NA} values
before the computation of the threshold}
}
\value{
A matrix with the transformed values.
}
\description{
This function transform input data via
a *Step Function*.
See \code{\link{stepFunction}} for further information.
}
\keyword{internal}
