% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{stepFunctionTranformationForVector}
\alias{stepFunctionTranformationForVector}
\title{Step Function Transformation for Vector}
\usage{
stepFunctionTranformationForVector(
  x,
  y = c(-1, 0, 1),
  thr,
  method = c("median", "mean", "mode", "midrange", "trimean", "iqm", "iqr", "mad"),
  na.rm = T
)
}
\arguments{
\item{x}{a (named) numerical vector}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical value, the threshold to use to
divide values in \code{x} in two intervals}

\item{method}{character string, indicating the score to be used as
threshold. This parameter is only used when \code{thr = NULL} or
not provided}

\item{na.rm}{logical, whether to remove \code{NA} values
before the computation of the threshold}
}
\value{
A vector with the transformed values.
}
\description{
This function transform input data via
a \emph{Step Function}.
See \code{\link{stepFunction}} for further information.
}
\author{
Alessandro Barberis
}
\keyword{internal}
