% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{subsetData}
\alias{subsetData}
\title{Subset Data}
\usage{
subsetData(x, i, rm.dupli = TRUE, ...)
}
\arguments{
\item{x}{(named) numerical vector or a matrix features-by-samples}

\item{i}{numerical vector giving the (row) position in \code{x}
or character vector matching the (row) names in \code{x}}

\item{rm.dupli}{logical, whether to remove
duplicated elements in \code{i}}
}
\value{
The subset data. It returns \code{NULL}
if \code{i} values are not present in \code{x}.
}
\description{
This function subset \code{x} given a
vector of values \code{i}. It is used to have uniform
behavior.
}
\examples{
\dontrun{

#vector
x = setNames(object = c(1,2,3,4,5), nm = LETTERS[1:5])
subsetData(x, i = c(1,3))
subsetData(x, i = LETTERS[c(1,5)])
subsetData(x, i = 20)

#matrix
x = matrix(
  data = sample(x = 1:100, size = 15),
  nrow = 5,
  ncol = 3,
  dimnames = list(LETTERS[1:5], letters[1:3])
)
subsetData(x, i = c(1,3))
subsetData(x, i = LETTERS[c(1,5)])
subsetData(x, i = 20)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
