% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{subsetMatrix}
\alias{subsetMatrix}
\title{Subset Data Matrix}
\usage{
subsetMatrix(x, i, rm.dupli = TRUE)
}
\arguments{
\item{x}{numerical matrix}

\item{i}{numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}}

\item{rm.dupli}{logical, whether to remove
duplicated elements in \code{i}. Useful argument when
computing random signatures}
}
\value{
The subset matrix. It returns \code{NULL}
if the subset is empty.
}
\description{
This function subset \code{x} given a
vector of values \code{i}. It is used to have uniform
behavior.
}
\examples{
\dontrun{
x = matrix(
  data = sample(x = 1:100, size = 15),
  nrow = 5,
  ncol = 3,
  dimnames = list(LETTERS[1:5], letters[1:3])
)
subsetMatrix(x, i = c(1,3))
subsetMatrix(x, i = LETTERS[c(1,5)])
subsetMatrix(x, i = 20)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
