% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{subsetVector}
\alias{subsetVector}
\title{Subset Data Vector}
\usage{
subsetVector(x, i, rm.dupli = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}}

\item{rm.dupli}{logical, whether to remove
duplicated elements in \code{i}}
}
\value{
The subset vector. It returns \code{NULL}
if \code{i} values are not present in \code{x}.
}
\description{
This function subset \code{x} given a
vector of values \code{i}. It is used to have uniform
behavior.
}
\examples{
\dontrun{
x = setNames(object = c(1,2,3,4,5), nm = LETTERS[1:5])
subsetVector(x, i = c(1,3))
subsetVector(x, i = LETTERS[c(1,5)])
subsetVector(x, i = 20)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
