% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{summation}
\alias{summation}
\title{Sum}
\usage{
summation(x, i = NULL, na.rm = T, ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[base]{sum}}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the sum of all
the values in \code{x}.
}
\details{
It is a wrapper to \code{\link[base]{sum}}
function.
}
\examples{
summation(x = c(1,2))

}
\author{
Alessandro Barberis
}
