% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{transformData}
\alias{transformData}
\title{Transform Data}
\usage{
transformData(x, f = c("none", "stepFunction", "quantile", "z"), ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{f}{string, the transformation method.
Available options are:
\describe{
\item{\code{'none'}}{\code{x} is returned}
\item{\code{'stepFunction'}}{step function}
\item{\code{'quantile'}}{quantile normalisation}
\item{\code{'z'}}{z-standardisation}
}}

\item{...}{further arguments to the transformation method.}
}
\value{
A numerical matrix containing the transformed
values.
}
\description{
This function is a dispatcher for the
selected transformation method.
}
\seealso{
\code{\link{stepFunctionTranformation}},
\code{\link{quantileNormalization}}
\code{\link{zStandardisation}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
