% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{trimean}
\alias{trimean}
\title{Tukey's Trimean}
\usage{
trimean(x, i = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
The \emph{trimean} contains information about the
center and some of the position of the data.
See the \strong{Details} section below for further information.
}
\details{
The trimean is defined as a weighted average of
the distribution's median and its two quartiles:

\deqn{TM(x) = \frac{Q_{1}(x) + 2Q_{2}(x) + Q_{3}(x)}{4} = \frac{median(x) + midhinge(x)}{2}}
}
\examples{
x = c(0,1,1,2,3,3,3,100,NA)
trimean(x = x, na.rm = TRUE)
}
\references{
https://en.wikipedia.org/wiki/Trimean
}
\author{
Alessandro Barberis
}
