% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{trimmedMeanScore}
\alias{trimmedMeanScore}
\title{Trimmed Arithmetic Mean Score}
\usage{
trimmedMeanScore(x, i, na.rm = TRUE, trim = 0)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
Trimmed Arithmetic Mean Score
}
\seealso{
[base::mean()]
}
\author{
Alessandro Barberis
}
