% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{trimmedMeanScores}
\alias{trimmedMeanScores}
\title{Trimmed Arithmetic Mean Scores}
\usage{
trimmedMeanScores(x, i, trim = 0, na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
Trimmed Arithmetic Mean Scores
}
\seealso{
\code{\link{trimmedMeanScore}},
\code{\link[base]{mean}}
}
\author{
Alessandro Barberis
}
