% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{updateSig}
\alias{updateSig}
\title{Subset Signature Vector}
\usage{
updateSig(x, i, rm.dupli = TRUE)
}
\arguments{
\item{x}{numerical vector or matrix}

\item{i}{numerical vector giving the rows (or positions)
in \code{x} or character vector matching the row
names (or names) in \code{x}}

\item{rm.dupli}{logical, whether to remove
duplicated elements in \code{i}}
}
\value{
The subset vector. If \code{updateSig}
removes cases, the indices of the cases form the
\code{"index"} attribute of the object.

It returns \code{NULL} if the subset is empty.
}
\description{
This function subset \code{i} given a
vector or matrix \code{x}. It is used to have uniform
behavior.
}
\examples{
\dontrun{

#x is a vector
x = setNames(
  object = c(1,2,3,4,5),
  nm = LETTERS[1:5]
)
updateSig(x, i = c(1,3))
updateSig(x, i = LETTERS[c(1,5)])
updateSig(x, i = LETTERS[c(1,6,5)])
updateSig(x, i = 20)

#x is a matrix
x = matrix(
  data = sample(x = 1:100, size = 15),
  nrow = 5,
  ncol = 3,
  dimnames = list(LETTERS[1:5], letters[1:3])
)
updateSig(x, i = c(1,3))
updateSig(x, i = LETTERS[c(1,5)])
updateSig(x, i = LETTERS[c(1,6,5)])
updateSig(x, i = 20)
}

}
\author{
Alessandro Barberis
}
\keyword{internal}
