% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{weightedMean}
\alias{weightedMean}
\title{Weighted Arithmetic Mean}
\usage{
weightedMean(x, w, i = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{w}{(optional) numerical vector of weights the same length as \code{x}}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link[stats]{weighted.mean}}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the *weighted arithmetic mean*.
}
\details{
It is a wrapper to \code{\link[stats]{weighted.mean}}
function.
}
\examples{
x = c(1,1,2,3,3,3,NA)
w = c(1,1,8,1,1,1,1)
weightedMean(x = x, w = w, na.rm = TRUE)

}
\author{
Alessandro Barberis
}
