% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{weightedSum}
\alias{weightedSum}
\title{Weighted Sum}
\usage{
weightedSum(x, w, i = NULL, na.rm = T)
}
\arguments{
\item{x}{(named) numerical vector}

\item{w}{numerical vector of weights the same length as
\code{x}}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes the sum of the
elements of a vector (\code{x}) multiplied by the elements
of another vector (\code{w}).
}
\examples{
x = c(1,1,2,3,3,3,NA)
w = c(1,1,8,1,1,1,1)
weightedSum(x = x, w = w, na.rm = TRUE)

}
\author{
Alessandro Barberis
}
