% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{weightedSumScore}
\alias{weightedSumScore}
\title{Weighted Sum Score}
\usage{
weightedSumScore(x, i, w, na.rm = TRUE, normalisation = "quantile", ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{w}{numerical vector of weights the same length as
\code{i}}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{normalisation}{the normalisation method to perform before
computing the weighted sum}

\item{...}{further arguments to \code{normalisation}}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes the weighted sum of
normalised input data \code{x}. It is computed as:

\deqn{wss(x) = \sum_{i=1}^{n}  xnorm_{i} * w_{i}}

where \eqn{n} is the number of \code{i} elements in \code{x};
\eqn{xnorm_{i}} is the \eqn{i}-th element of the normalised vector;
\eqn{w_{i}} is the \eqn{i}-th element of the vector of weights.
}
\details{
The input data \code{x} is firstly normalised using
the technique chosen via the \code{normalisation} parameter.
Then, each element of \code{i} in \code{x} is weighted by
the corresponding element in \code{w}. Finally, the sum of
the weighted elements is calculated.
}
\seealso{
\code{\link{normaliseData}}
}
\author{
Alessandro Barberis
}
