% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{weightedSumScores}
\alias{weightedSumScores}
\title{Weighted Sum Scores}
\usage{
weightedSumScores(x, i, w, na.rm = TRUE, normalisation = "quantile", ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{w}{numerical vector of weights the same length as
\code{i}}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{normalisation}{the normalisation method to perform before
computing the weighted sum}

\item{...}{further arguments to \code{normalisation}}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
Weighted Sum Scores
}
\examples{
\dontrun{
x = matrix(
data = sample(x = 1:100, size = 15),
nrow = 5,
ncol = 3,
dimnames = list(LETTERS[1:5], letters[1:3]))

weightedSumScores(x = x)

}
}
\seealso{
\code{\link{normaliseData}}
}
\author{
Alessandro Barberis
}
