% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{zscore}
\alias{zscore}
\title{Enrichment Score}
\usage{
zscore(...)
}
\arguments{
\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes an enrichment score.
}
\details{
This function always returns NA.
}
\examples{
zscore()
}
\author{
Alessandro Barberis
}
