% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{zStandardisation}
\alias{zStandardisation}
\title{Z-Standardisation}
\usage{
zStandardisation(x, robust = FALSE, by = c("rows", "cols"), na.rm = TRUE)
}
\arguments{
\item{x}{a (named) numerical vector or a matrix features-by-samples.}

\item{robust}{logical, whether to compute a robust z-standardisation.}

\item{by}{character string, indicating whether to compute the
standardisation by rows or columns of \code{x}.
It is used when \code{x} is a matrix.}

\item{na.rm}{logical, whether to remove \code{NA} values
before the computation.}
}
\value{
A numerical vector or matrix, containing the output
of the transformation.
}
\description{
This function transform the input
data via z-standardisation.
See the \strong{Details} section below for further information.
}
\details{
The z-standardisation (or z-score normalisation) is a method for
transforming the data so that it has a mean of zero and a standard deviation
of one. It is computed as:

\deqn{z(x) = \frac{x - \mu}{\sigma}}

where \eqn{\mu} and \eqn{\sigma} are the mean and standard deviation of the
population, respectively.

Since z-scores can be affected by unusually large or small data values,
we can also compute a more robust modified version as:

\deqn{z(x) = \frac{x - \mathrm{median}}{\mathrm{MAD}}}

where \eqn{\mathrm{MAD}} is the median absolute deviation of the population.
}
\examples{
#set seed for reproducibility
set.seed(seed = 5381L)

#x is a vector
x = sample(x = -10:10, size = 10)
#Standard z-score transformation
zStandardisation(x)
#Robust z-score transformation
zStandardisation(x=x,robust=TRUE)

#x is a matrix
#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)
#Standardise the features
zStandardisation(x=x,by="rows")

}
\seealso{
\code{\link{zStandardisationForVector}},
\code{\link{zStandardisationForMatrix}}
}
\author{
Alessandro Barberis
}
