% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{zStandardisationForVector}
\alias{zStandardisationForVector}
\title{Z-Standardisation for Vector}
\usage{
zStandardisationForVector(x, robust = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{a (named) numerical vector.}

\item{robust}{logical, whether to compute a robust z-standardisation.}

\item{na.rm}{logical, whether to remove \code{NA} values
before the computation.}
}
\value{
A vector with the transformed values.
}
\description{
This function transform the input
data via z-standardisation.
See the \strong{Details} section below for further information.
}
\details{
The z-standardisation (or z-score normalisation) is a method for
transforming the data so that it has a mean of zero and a standard deviation
of one. It is computed as:

\deqn{z(x) = \frac{x - \mu}{\sigma}}

where \eqn{\mu} and \eqn{\sigma} are the mean and standard deviation of the
population, respectively.

Since z-scores can be affected by unusually large or small data values,
we can also compute a more robust modified version as:

\deqn{z(x) = \frac{x - \mathrm{median}}{\mathrm{MAD}}}

where \eqn{\mathrm{MAD}} is the median absolute deviation of the population.
}
\references{
https://en.wikipedia.org/wiki/Standard_score
}
\author{
Alessandro Barberis
}
\keyword{internal}
