% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{bristowScores}
\alias{bristowScores}
\title{Bristow Scores}
\usage{
bristowScores(x, i, m = c("rows", "cols"), na.rm = TRUE)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{m}{character string, indicating whether to
compute the median abundance across the rows
of \code{x} (\code{m = "rows"}) or across the
columns of \code{x} (\code{m = "cols"}).
In the first case, the median abundance is the same
across signature elements within the same sample.
In the second case, the median abundance of the
\eqn{i}-th element of the signature is the same across
samples, and different elements of \code{i} have
different median abundances}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
This function computes the Bristow score
from each column vector in the input matrix.
See \code{\link{bristowScore}} for further details.
}
\references{
Lalonde, E. et al., \emph{Tumour genomic and microenvironmental heterogeneity for integrated prediction of 5-year biochemical recurrence of prostate cancer: a retrospective cohort study},
The Lancet Oncology (2014), appendix
}
\seealso{
\code{\link{bristowScore}}
}
\author{
Alessandro Barberis
}
