% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{colWeightedSums}
\alias{colWeightedSums}
\title{Column Weighted Sums}
\usage{
colWeightedSums(x, w, rows = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{\link{weightedSum}}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function computes the \emph{weighted sum}
for each column in an input matrix.
}
\examples{
x = matrix(data = c(1,3,1,4), ncol = 2)
w = c(5, 1)
colWeightedSums(x = x, w = w)
}
\author{
Alessandro Barberis
}
