% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-significance.R
\name{computeCI}
\alias{computeCI}
\title{Confidence interval of estimate}
\usage{
computeCI(obs, rnd, conf.level = 0.95, type = c("standard", "percentile"), ...)
}
\arguments{
\item{obs}{numerical vector containing the observed value of
the test statistic}

\item{rnd}{numerical vector containing the permutation/bootstrap
replication of the statistics}

\item{conf.level}{the desired confidence level}

\item{type}{character string indicating which formula to use
in the computation of the confidence interval. Available options are:
\describe{
  \item{\code{"standard"}}{the standard confidence interval}
  \item{\code{"percentile"}}{the percentile confidence interval}
}}
}
\value{
A named numeric vector with 2 elements, \code{lower} and \code{upper},
the lower and upper bounds of the confidence interval
for the point estimate.
}
\description{
This function computes an approximate confidence interval of
a point estimate (i.e. the test statistic).
}
\seealso{
\code{\link{computeStandardCi}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
