% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{computeColMeasures}
\alias{computeColMeasures}
\title{Computes Measure for Each Column in a Matrix}
\usage{
computeColMeasures(
  x,
  rows = NULL,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqm", "iqr", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{a numerical matrix features-by-samples}

\item{rows}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x} to operate over.
If \code{missing} or \code{rows = NULL}, all the rows
in \code{x} are considered for the computation of
the measures}

\item{score}{character string indicating the statistical measure
to compute}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical vector containing the computed
score for each column in \code{x}.
}
\description{
This function calculates statistical measure(s)
for each column in an input matrix.
It is an internal function, and it is not intended to be
called by users.
}
\details{
Internally, \code{\link{computeColMeasures}} uses these
functions to compute the measures:

\describe{
\item{\code{"sum"         }}{\code{\link{colSummations}}}
\item{\code{"weightedSum" }}{\code{\link{colWeightedSums}}}
\item{\code{"mean"        }}{\code{\link{colArithmeticMeans}}}
\item{\code{"trimmedMean" }}{\code{\link{colTrimmedMeans}}}
\item{\code{"weightedMean"}}{\code{\link{colWeightedArithmeticMeans}}}
\item{\code{"median"      }}{\code{\link{colMidpoints}}}
\item{\code{"mode"        }}{\code{\link{colModes}}}
\item{\code{"midrange"    }}{\code{\link{colMidranges}}}
\item{\code{"midhinge"    }}{\code{\link{colMidhinges}}}
\item{\code{"trimean"     }}{\code{\link{colTrimeans}}}
\item{\code{"iqr"         }}{\code{\link{colIQRs}}}
\item{\code{"iqm"         }}{\code{\link{colIQMs}}}
\item{\code{"mad"         }}{\code{\link{colMADs}}}
\item{\code{"aad"         }}{\code{\link{colAADs}}}
\item{\code{"ssgsea"      }}{\code{\link{colSsgsea}}}
\item{\code{"gsva"        }}{\code{\link{colGsva}}}
\item{\code{"plage"       }}{\code{\link{colPlage}}}
\item{\code{"zscore"      }}{\code{\link{colZscore}}}
}
Look at the different functions to know which specific
arguments they accept (arguments can be passed via the
\code{...} parameter).
}
\author{
Alessandro Barberis
}
\keyword{internal}
