% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-significance.R
\name{computeSignificance}
\alias{computeSignificance}
\title{Compute the Significance of the Test}
\usage{
computeSignificance(
  obs,
  rnd,
  type = c("asl", "ci", "se"),
  ci = c("standard", "percentile"),
  conf.level = 0.95
)
}
\arguments{
\item{obs}{numerical vector containing the test statistic}

\item{rnd}{numerical vector containing the permutation/bootstrap
statistics}

\item{type}{character string indicating the significance to
compute. Three options are available:
\describe{
\item{\code{asl}}{achieved significance level}
\item{\code{ci}}{confidence interval}
\item{\code{se}}{standard error}
}}

\item{ci}{character string indicating the type of
confidence interval. Two options are available:
\describe{
\item{\code{standard}}{the standard confidence interval}
\item{\code{percentile}}{the percentile confidence interval}
}}
}
\description{
Compute the Significance of the Test
}
\seealso{
\code{\link{computeASL}},
\code{\link{computeCI}},
\code{\link{computeSE}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
