% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-significance.R
\name{computeStandardCi}
\alias{computeStandardCi}
\title{Standard Confidence Interval}
\usage{
computeStandardCi(
  obs,
  rnd,
  conf.level = 0.95,
  distribution = c("normal", "t"),
  n
)
}
\arguments{
\item{obs}{numerical vector containing the observed value of
the test statistic}

\item{rnd}{numerical vector containing the permutation/bootstrap
replication of the statistics}

\item{conf.level}{the desired confidence level}

\item{distribution}{sampling distribution of the estimate.
Use \code{normal} if the population has unknown mean and known variance (or if \code{n} is large),
\code{t} if population has unknown mean and variance}

\item{n}{sample size, used to compute the degrees of freedom if \code{distribution = "t"}}
}
\value{
A named numeric vector with 2 elements, \code{lower} and \code{upper},
the lower and upper bounds of the confidence interval
for the point estimate.
}
\description{
This function computes an approximate confidence interval of
a point estimate (i.e. the test statistic).
}
\details{
The confidence interval provides additional information about the
variability of a point estimate and it is generally defined as

\deqn{CI = estimate \pm \textit{margin of error} = estimate \pm \textit{critical value} \times \textit{standard error}}

where \eqn{\textit{estimate}} is the sample statistic estimating the population parameter of interest;
\eqn{\textit{critical value}} is a value based on the sampling distribution of the estimate
and the desired confidence level; \eqn{\textit{standard error}} is the standard deviation of the
point estimate.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5723800/}
}
\author{
Alessandro Barberis
}
\keyword{internal}
