% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment Score}
\usage{
enrichment(method = c("ssgsea", "gsva", "plage", "zscore"), ...)
}
\arguments{
\item{method}{character string, the name of the method
to be used for computing the enrichment scores.
Current available options are:
\describe{
\item{\code{'ssgsea'}}{Single Sample Gene Set Enrichment Analysis}
\item{\code{'gsva'}}{Gene Set Variation Analysis}
\item{\code{'plage'}}{Pathway Level Analysis of Gene Expression}
\item{\code{'zscore'}}{Z-Score}
}}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed measure.
A default \code{NA} value is returned if the score can't be
computed.
}
\description{
This function computes an enrichment score.
}
\details{
This function always returns NA.
}
\examples{
enrichment()

}
\author{
Alessandro Barberis
}
\keyword{internal}
