% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{genericScorer}
\alias{genericScorer}
\title{Generic Scorer Function}
\usage{
genericScorer(
  x,
  i = NULL,
  w = NULL,
  na.rm = T,
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore"),
  transform.fun = NULL,
  transform.args = list(),
  transform.sub = FALSE,
  ...
)
}
\arguments{
\item{x}{(named) numerical vector or matrix}

\item{i}{(optional) numerical vector giving the (row) position
in \code{x} or character vector matching the (row) names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score}

\item{w}{numerical vector of weights the same length as
\code{i}. It is used for the computation of the weighted scores}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{score}{character string indicating the summary score to compute}

\item{transform.fun}{function to transform the data.
If provided, \code{x} is transformed
(\code{x = transform.fun(x, transform.args)})
before the computation of the scores.
See \code{\link{getDataTransformer}} for further details
about built-in options}

\item{transform.args}{list of parameters to the
data transformation function}

\item{transform.sub}{logical, whether to transform
\code{x} after it is subset for \code{i} (used to
speedup computation). Default is \code{FALSE},
meaning the transformation would be applied directly
to \code{x} provided in input}

\item{...}{further arguments to \code{score}}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
This function computes a summary score
from an input vector or from each column vector
of an input matrix.
}
\details{
The function handles input matrices by calling
\code{\link{computeColMeasures}}.
Vectors are converted to 1 column matrices.

Internally, \code{\link{computeColMeasures}} uses these
functions to compute the measures:
\describe{
\item{\code{"sum"         }}{\code{\link{colSummations}}}
\item{\code{"weightedSum" }}{\code{\link{colWeightedSums}}}
\item{\code{"mean"        }}{\code{\link{colArithmeticMeans}}}
\item{\code{"trimmedMean" }}{\code{\link{colTrimmedMeans}}}
\item{\code{"weightedMean"}}{\code{\link{colWeightedArithmeticMeans}}}
\item{\code{"median"      }}{\code{\link{colMidpoints}}}
\item{\code{"mode"        }}{\code{\link{colModes}}}
\item{\code{"midrange"    }}{\code{\link{colMidranges}}}
\item{\code{"midhinge"    }}{\code{\link{colMidhinges}}}
\item{\code{"trimean"     }}{\code{\link{colTrimeans}}}
\item{\code{"iqr"         }}{\code{\link{colIQRs}}}
\item{\code{"iqm"         }}{\code{\link{colIQMs}}}
\item{\code{"mad"         }}{\code{\link{colMADs}}}
\item{\code{"aad"         }}{\code{\link{colAADs}}}
\item{\code{"ssgsea"      }}{\code{\link{colSsgsea}}}
\item{\code{"gsva"        }}{\code{\link{colGsva}}}
\item{\code{"plage"       }}{\code{\link{colPlage}}}
\item{\code{"zscore"      }}{\code{\link{colZscore}}}
}
}
\seealso{
\code{\link{getDataTransformer}},
\code{\link{computeColMeasures}}
}
\author{
Alessandro Barberis
}
\keyword{internal}
