% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-scorers.R
\name{getScorer}
\alias{getScorer}
\title{Get Scorer}
\usage{
getScorer(
  score = c("sum", "weightedSum", "mean", "trimmedMean", "weightedMean", "median",
    "mode", "midrange", "midhinge", "trimean", "iqr", "iqm", "mad", "aad", "ssgsea",
    "gsva", "plage", "zscore")
)
}
\arguments{
\item{score}{character string, one of the supported measures}
}
\value{
A function.
}
\description{
This function is a dispatcher for the score
function selected in input.
}
\author{
Alessandro Barberis
}
