% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-data-preprocessing.R
\name{quantileNormalization}
\alias{quantileNormalization}
\title{Quantile Normalization}
\usage{
quantileNormalization(
  x,
  na.rm = TRUE,
  ties.method = c("min", "max", "first", "last")
)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{ties.method}{a character string specifying how ties are treated,
    see \sQuote{Details}; can be abbreviated.}
}
\value{
A numerical matrix containing the normalized
values.
}
\description{
This function normalises the input
data via quantile normalisation technique.
}
\examples{
\dontrun{

x = as.data.frame(
matrix(
data = c(2,3,6,5,5,3,6,6,
4,3,5,3,3,5,4,5,
5,4,3,4,7,2,5,5,
3,5,4,4,5,6,3,4,
4,5,5,6,6,5,5,7),
nrow = 5,
byrow = TRUE))

quantileNormalization(x)

}

}
\references{
https://www.nature.com/articles/s41598-020-72664-6,
https://davetang.org/muse/2014/07/07/quantile-normalisation-in-r/
}
\author{
Alessandro Barberis
}
\keyword{internal}
