% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{randomSignatures}
\alias{randomSignatures}
\title{Random Sampling}
\usage{
randomSignatures(x, i, n = 1L, exclude = T)
}
\arguments{
\item{x}{a numerical vector or matrix}

\item{i}{(optional) numerical vector giving the (row)
elements in \code{x} or character vector matching the
(row) names in \code{x}}

\item{n}{integer, number of repeated samples
to generate}

\item{exclude}{logical, whether to exclude
\code{i} from the possible values}
}
\value{
A list of length \code{n} where each element
is a vector of indices representing the sampled data.
}
\description{
This function takes \code{n} samples of
size \code{length(i)} from the elements of \code{x}.
Two types of sampling are available: using all
elements of \code{x} or using the elements of \code{x}
that are not part of \code{i}.
}
\examples{
#set seed for reproducibility
set.seed(seed = 5381L)

#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Set signature
i = rownames(x)[1:10]

#Generate 1 random signatures
#(elements of i are possible)
randomSignatures(
 x = x,
 i = i,
 n = 1,
 exclude = F
)

#Generate 1 random signatures
#(elements of i are excluded)
randomSignatures(
 x = x,
 i = i,
 n = 1,
 exclude = T
)

}
\seealso{
\code{\link[base]{sample}}
}
\author{
Alessandro Barberis
}
