% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utility-functions.R
\name{sampleData}
\alias{sampleData}
\title{Random Sampling}
\usage{
sampleData(x, method = c("none", "permutation", "bootstrap"), n = 1L)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{method}{string, the sampling method to use.
Available options are:

Three options are available:
\describe{
\item{\code{none}}{a sequence from 1 to \code{nrow(x)} is returned}
\item{\code{permutation}}{random sampling without replacement}
\item{\code{bootstrap}}{random sampling with replacement}
}}

\item{n}{integer, number of repeated samples
to generate}
}
\value{
A list of length \code{n} where each element
is a vector of indices representing the sampled data.
}
\description{
This function takes \code{n} samples of
size \code{nrow(x)} from the elements of \code{x}.
Two types of sampling are available: sampling with
or without replacement.
}
\examples{
#set seed for reproducibility
set.seed(seed = 5381L)

#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)

#Take one sample via permutation
#(duplicated elements are not possible)
sampleData(
 x = x,
 method = 'permutation',
 n = 1
)

#Take one sample via bootstrap
#(duplicated elements are possible)
sampleData(
 x = x,
 method = 'bootstrap',
 n = 1
)

}
\seealso{
\code{\link[base]{sample}}
}
\author{
Alessandro Barberis
}
