% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-stats-functions.R
\name{stepFunction}
\alias{stepFunction}
\title{Step Function}
\usage{
stepFunction(x, y = c(-1, 0, 1), thr = 0, by = c("rows", "cols"))
}
\arguments{
\item{x}{a (named) numerical vector or a matrix features-by-samples}

\item{y}{numerical vector of length 3, the values to
use for the piecewise function.
The first value is used for the left interval,
the second value is used when \code{x == threshold},
and the third value is used for the right interval}

\item{thr}{numerical value, the threshold to use to
divide values in \code{x} in two intervals}

\item{by}{character string, indicating whether to
apply the threshold to rows or columns of
\code{x}.
In the first case, the threshold for the \eqn{i}-th row
is the same across columns.
In the second case, the threshold is the same within
the same column vector.
It is used when \code{x} is a matrix}
}
\value{
A numerical vector or matrix, containing the output
of the step function for each element of \code{x}.
The default settings of \code{thr} and \code{y}
makes this function equivalent to a sign function.
}
\description{
This function implements a *Step Function*.
See the **Details** section below for further information.
}
\details{
This function implements a piecewise function
that is defined as follows:

\deqn{SF = f(x) =
\left\{
\begin{array}{l}
y_{1}, \quad \textrm{if} \quad x > threshold \\
y_{2}, \quad \textrm{if} \quad x = threshold \\
y_{3}, \quad \textrm{if} \quad x < threshold
\end{array}
\right\}
}
}
\examples{
#set seed for reproducibility
set.seed(seed = 5381L)

#x is a vector
x = sample(x = -10:10, size = 10)
stepFunction(x, thr = 0)

#x is a matrix
#Define row/col size
nr = 20
nc = 10

#Create input matrix
x = matrix(
 data = stats::runif(n = nr*nc, min = 0, max = 1000),
 nrow = nr,
 ncol = nc,
 dimnames = list(
   paste0("g",seq(nr)),
   paste0("S",seq(nc))
 )
)
#compute
stepFunction(x, thr = 500)

}
\author{
Alessandro Barberis
}
