% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-vector-scores.R
\name{weightedSumScore}
\alias{weightedSumScore}
\title{Weighted Sum Score}
\usage{
weightedSumScore(x, i, w, na.rm = TRUE, normalisation = "quantile", ...)
}
\arguments{
\item{x}{(named) numerical vector}

\item{i}{(optional) numerical vector giving the position in \code{x}
or character vector matching the names in \code{x}.
If \code{missing} or \code{i = NULL}, the entire \code{x} is
considered for the computation of the score.}

\item{w}{numerical vector of weights the same length as
\code{i}}

\item{na.rm}{logical, whether to remove \code{NA}
values from \code{x} before computation}

\item{normalisation}{the normalisation method to perform before
computing the weighted sum}

\item{...}{further arguments to \code{normalisation}}
}
\value{
A numerical value representing the computed score.
A default \code{NA} value is returned if the score can't be
computed, or if \code{i} values are not present in \code{x}.
}
\description{
Weighted Sum Score
}
\seealso{
\code{\link[stats]{weighted.mean}},
\code{\link{quantileNormalization}}
}
\author{
Alessandro Barberis
}
