% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-matrix-scores.R
\name{weightedSumScores}
\alias{weightedSumScores}
\title{Weighted Sum Scores}
\usage{
weightedSumScores(x, i, w, na.rm = TRUE, normalisation = "quantile", ...)
}
\arguments{
\item{x}{numerical matrix, features-by-samples}

\item{i}{(optional) numerical vector giving the rows
in \code{x} or character vector matching the row
names in \code{x}.
If \code{missing} or \code{i = NULL}, all the rows
in \code{x} are considered for the computation of
the score}

\item{w}{numerical vector of weights the same length as
\code{i}}

\item{na.rm}{logical, whether to remove \code{NA}
values before computation}

\item{normalisation}{the normalisation method to perform before
computing the weighted sum}

\item{...}{further arguments to \code{normalisation}}
}
\value{
A numerical vector containing the computed
score for each sample.
}
\description{
This function computes the weighted sum
from each column vector in the normalised input matrix.
The score for the \eqn{j}-th sample is computed as:

\deqn{wss_{j} = \sum_{i=1}^{n}  xnorm_{ij} * w_{i}}

where \eqn{n} is the number of \code{i} elements in \code{x};
\eqn{xnorm_{ij}} is the \eqn{i}-th row for the \eqn{j}-th column
of the normalised matrix;
\eqn{w_{i}} is the \eqn{i}-th element of weights vector.
}
\details{
The input data \code{x} is firstly normalised using
the technique chosen via the \code{normalisation} parameter.
Then, for each column of \code{x}, each element of \code{i}
in \code{x} is weighted by the corresponding element in \code{w}.
Finally, the sum of the weighted elements is calculated.
}
\examples{
\dontrun{
x = matrix(
data = sample(x = 1:100, size = 15),
nrow = 5,
ncol = 3,
dimnames = list(LETTERS[1:5], letters[1:3]))

weightedSumScores(x = x)

}
}
\seealso{
\code{\link{normaliseData}}
}
\author{
Alessandro Barberis
}
